+++ packages/bluesky/lib/src/services/unspecced_service.dart
@@ -48,6 +48,7 @@ final class UnspeccedService {
   /// https://atprotodart.com/docs/lexicons/app/bsky/unspecced/searchActorsSkeleton
   Future<core.XRPCResponse<SkeletonActorsByQuery>> searchActorsSkeleton(
     final String query, {
+    String? viewer,
     bool? typeahead,
     int? limit,
     String? cursor,
@@ -55,6 +56,7 @@ final class UnspeccedService {
       // ignore: deprecated_member_use_from_same_package
       await searchActorsByQuerySkeleton(
         query,
+        viewer: viewer,
         typeahead: typeahead,
         limit: limit,
         cursor: cursor,
@@ -96,12 +98,14 @@ final class UnspeccedService {
   @Deprecated('Use .searchActorsSkeleton instead. Will be removed')
   Future<core.XRPCResponse<SkeletonActorsByQuery>> searchActorsByQuerySkeleton(
     final String query, {
+    String? viewer,
     bool? typeahead,
     int? limit,
     String? cursor,
   }) async =>
       await _searchActorsByQuerySkeleton(
         query: query,
+        viewer: viewer,
         typeahead: typeahead,
         limit: limit,
         cursor: cursor,
@@ -142,6 +146,7 @@ final class UnspeccedService {
 
   Future<core.XRPCResponse<T>> _searchActorsByQuerySkeleton<T>({
     required String query,
+    required String? viewer,
     required bool? typeahead,
     required int? limit,
     required String? cursor,
@@ -151,6 +156,7 @@ final class UnspeccedService {
         ns.appBskyUnspeccedSearchActorsSkeleton,
         parameters: _buildSearchActorsSkeletonParams(
           query: query,
+          viewer: viewer,
           typeahead: typeahead,
           limit: limit,
           cursor: cursor,
@@ -182,12 +188,14 @@ final class UnspeccedService {
 
   Map<String, dynamic> _buildSearchActorsSkeletonParams({
     required String query,
+    required String? viewer,
     required bool? typeahead,
     required int? limit,
     required String? cursor,
   }) =>
       {
         'q': query,
+        'viewer': viewer,
         'typeahead': typeahead,
         'limit': limit,
         'cursor': cursor,